Fixed problems with vector linking. Fix 3rd vector component factor not applied when the var selected is the 3rdd component. Fixed value extraction for 3 component vector field to use the 3 components for speed instead of only 2. Fixed graphitem size not taken into account when using double notation.

+++ LibTkGL/TkViewport/tkGraphItem.h
@@ -60,7 +60,7 @@ typedef struct TGraphItem {
 
    double         Origin;                     /*Origine du graph selon l'axe et l'orientation*/
    int            Width;                      /*line width*/
-   int            Size;                       /*Icon Size*/
+   double         Size;                       /*Icon Size*/
    char          *Orient;                     /*Orientation (X,Y,Z,XY,YZ,XZ)*/
    char          *Fit;                        /*Fitting type*/
    int            Icon;                       /*Icone*/
